Extract SecureToken helper for shared random/hash logic
